// packages/langium-crmscript/src/generated/ast.ts
export interface BinaryExpression extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'BinaryExpression';
    left: Expression;
    operator: '!=' | '*' | '+' | '-' | '/' | '<' | '<=' | '=' | '==' | '>' | '>=' | 'and' | 'or';
    right: Expression;
}

export const BinaryExpression = {
    $type: 'BinaryExpression',
    left: 'left',
    operator: 'operator',
    right: 'right'
} as const;

export function isBinaryExpression(item: unknown): item is BinaryExpression {
    return reflection.isInstance(item, BinaryExpression.$type);
}

export interface BooleanExpression extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'BooleanExpression';
    value: boolean;
}

export const BooleanExpression = {
    $type: 'BooleanExpression',
    value: 'value'
} as const;

export function isBooleanExpression(item: unknown): item is BooleanExpression {
    return reflection.isInstance(item, BooleanExpression.$type);
}

export interface Class extends langium.AstNode {
    readonly $container: DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'Class';
    constructors: Array<Constructor>;
    members: Array<ClassMember>;
    name: string;
}

export const Class = {
    $type: 'Class',
    constructors: 'constructors',
    members: 'members',
    name: 'name'
} as const;

export function isClass(item: unknown): item is Class {
    return reflection.isInstance(item, Class.$type);
}

export type ClassMember = FieldMember | MethodMember;

export const ClassMember = {
    $type: 'ClassMember'
} as const;

export function isClassMember(item: unknown): item is ClassMember {
    return reflection.isInstance(item, ClassMember.$type);
}

export interface Constructor extends langium.AstNode {
    readonly $container: Class;
    readonly $type: 'Constructor';
    parameters: Array<Parameter>;
}

export const Constructor = {
    $type: 'Constructor',
    parameters: 'parameters'
} as const;

export function isConstructor(item: unknown): item is Constructor {
    return reflection.isInstance(item, Constructor.$type);
}

export interface ConstructorCall extends langium.AstNode {
    readonly $container: VariableDeclaration;
    readonly $type: 'ConstructorCall';
    params: Array<Parameter>;
    type: langium.Reference<Class>;
}

export const ConstructorCall = {
    $type: 'ConstructorCall',
    params: 'params',
    type: 'type'
} as const;

export function isConstructorCall(item: unknown): item is ConstructorCall {
    return reflection.isInstance(item, ConstructorCall.$type);
}

export type DefinitionElement = Globals | NamedElement;

export const DefinitionElement = {
    $type: 'DefinitionElement'
} as const;

export function isDefinitionElement(item: unknown): item is DefinitionElement {
    return reflection.isInstance(item, DefinitionElement.$type);
}

export interface DefinitionModel extends langium.AstNode {
    readonly $type: 'DefinitionModel';
    definitionelements: Array<DefinitionElement>;
}

export const DefinitionModel = {
    $type: 'DefinitionModel',
    definitionelements: 'definitionelements'
} as const;

export function isDefinitionModel(item: unknown): item is DefinitionModel {
    return reflection.isInstance(item, DefinitionModel.$type);
}

export interface Enum extends langium.AstNode {
    readonly $container: DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'Enum';
    enumMembers: Array<EnumMember>;
    name: string;
}

export const Enum = {
    $type: 'Enum',
    enumMembers: 'enumMembers',
    name: 'name'
} as const;

export function isEnum(item: unknown): item is Enum {
    return reflection.isInstance(item, Enum.$type);
}

export interface EnumMember extends langium.AstNode {
    readonly $container: DefinitionModel | Enum | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'EnumMember';
    name: string;
    value: number;
}

export const EnumMember = {
    $type: 'EnumMember',
    name: 'name',
    value: 'value'
} as const;

export function isEnumMember(item: unknown): item is EnumMember {
    return reflection.isInstance(item, EnumMember.$type);
}

export type Expression = BinaryExpression | BooleanExpression | MemberCall | NilExpression | NumberExpression | StringExpression | UnaryExpression;

export const Expression = {
    $type: 'Expression'
} as const;

export function isExpression(item: unknown): item is Expression {
    return reflection.isInstance(item, Expression.$type);
}

export interface ExpressionBlock extends langium.AstNode {
    readonly $container: ExpressionBlock | ForStatement | FunctionDeclaration | IfStatement | ImplementationModel | WhileStatement;
    readonly $type: 'ExpressionBlock';
    statements: Array<Statement>;
    types: Array<Type>;
}

export const ExpressionBlock = {
    $type: 'ExpressionBlock',
    statements: 'statements',
    types: 'types'
} as const;

export function isExpressionBlock(item: unknown): item is ExpressionBlock {
    return reflection.isInstance(item, ExpressionBlock.$type);
}

export interface FieldMember extends langium.AstNode {
    readonly $container: Class | DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'FieldMember';
    name: string;
    type: langium.Reference<Class>;
}

export const FieldMember = {
    $type: 'FieldMember',
    name: 'name',
    type: 'type'
} as const;

export function isFieldMember(item: unknown): item is FieldMember {
    return reflection.isInstance(item, FieldMember.$type);
}

export interface ForExecution extends langium.AstNode {
    readonly $container: ForStatement;
    readonly $type: 'BinaryExpression' | 'BooleanExpression' | 'Expression' | 'ForExecution' | 'MemberCall' | 'NilExpression' | 'NumberExpression' | 'StringExpression' | 'UnaryExpression';
    type: langium.Reference<VariableDeclaration>;
}

export const ForExecution = {
    $type: 'ForExecution',
    type: 'type'
} as const;

export function isForExecution(item: unknown): item is ForExecution {
    return reflection.isInstance(item, ForExecution.$type);
}

export interface ForStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'ForStatement';
    block: ExpressionBlock;
    condition?: Expression;
    counter?: NamedElement;
    execution?: ForExecution;
}

export const ForStatement = {
    $type: 'ForStatement',
    block: 'block',
    condition: 'condition',
    counter: 'counter',
    execution: 'execution'
} as const;

export function isForStatement(item: unknown): item is ForStatement {
    return reflection.isInstance(item, ForStatement.$type);
}

export interface FunctionDeclaration extends langium.AstNode {
    readonly $container: DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'FunctionDeclaration';
    array: boolean;
    arrayOfArray: boolean;
    body: ExpressionBlock;
    name: string;
    parameters: Array<Parameter>;
    returnType: langium.Reference<Class>;
}

export const FunctionDeclaration = {
    $type: 'FunctionDeclaration',
    array: 'array',
    arrayOfArray: 'arrayOfArray',
    body: 'body',
    name: 'name',
    parameters: 'parameters',
    returnType: 'returnType'
} as const;
// ...
export interface IfStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'IfStatement';
    block: ExpressionBlock;
    condition: Expression;
    elseBlock?: ExpressionBlock;
}

export const IfStatement = {
    $type: 'IfStatement',
    block: 'block',
    condition: 'condition',
    elseBlock: 'elseBlock'
} as const;

export function isIfStatement(item: unknown): item is IfStatement {
    return reflection.isInstance(item, IfStatement.$type);
}

export interface ImplementationModel extends langium.AstNode {
    readonly $type: 'ImplementationModel';
    includes: Array<Include>;
    statements: Array<Statement>;
    types: Array<Type>;
}

export const ImplementationModel = {
    $type: 'ImplementationModel',
    includes: 'includes',
    statements: 'statements',
    types: 'types'
} as const;

export function isImplementationModel(item: unknown): item is ImplementationModel {
    return reflection.isInstance(item, ImplementationModel.$type);
}

export interface Include extends langium.AstNode {
    readonly $container: ImplementationModel;
    readonly $type: 'Include';
    file: string;
}

export const Include = {
    $type: 'Include',
    file: 'file'
} as const;

export function isInclude(item: unknown): item is Include {
    return reflection.isInstance(item, Include.$type);
}

export interface MemberCall extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'MemberCall';
    arguments: Array<Expression>;
    element?: langium.Reference<NamedElement>;
    explicitOperationCall: boolean;
    previous?: Expression;
}

export const MemberCall = {
    $type: 'MemberCall',
    arguments: 'arguments',
    element: 'element',
    explicitOperationCall: 'explicitOperationCall',
    previous: 'previous'
} as const;

export function isMemberCall(item: unknown): item is MemberCall {
    return reflection.isInstance(item, MemberCall.$type);
}

export interface MethodMember extends langium.AstNode {
    readonly $container: Class | DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'MethodMember';
    array: boolean;
    arrayOfArray: boolean;
    name: string;
    parameters: Array<Parameter>;
    returnType: langium.Reference<Class>;
}

export const MethodMember = {
    $type: 'MethodMember',
    array: 'array',
    arrayOfArray: 'arrayOfArray',
    name: 'name',
    parameters: 'parameters',
    returnType: 'returnType'
} as const;

export function isMethodMember(item: unknown): item is MethodMember {
    return reflection.isInstance(item, MethodMember.$type);
}

export type NamedElement = Class | Enum | EnumMember | FieldMember | FunctionDeclaration | MethodMember | VariableDeclaration;

export const NamedElement = {
    $type: 'NamedElement'
} as const;
// ...
export interface NumberExpression extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'NumberExpression';
    value: number;
}

export const NumberExpression = {
    $type: 'NumberExpression',
    value: 'value'
} as const;

export function isNumberExpression(item: unknown): item is NumberExpression {
    return reflection.isInstance(item, NumberExpression.$type);
}

export interface Parameter extends langium.AstNode {
    readonly $container: Constructor | ConstructorCall | FunctionDeclaration | Globals | MethodMember;
    readonly $type: 'Parameter';
    array: boolean;
    arrayOfArray: boolean;
    name: string;
    type: langium.Reference<Class>;
}

export const Parameter = {
    $type: 'Parameter',
    array: 'array',
    arrayOfArray: 'arrayOfArray',
    name: 'name',
    type: 'type'
} as const;

export function isParameter(item: unknown): item is Parameter {
    return reflection.isInstance(item, Parameter.$type);
}

export interface PrintStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'PrintStatement';
    value: StringExpression;
}

export const PrintStatement = {
    $type: 'PrintStatement',
    value: 'value'
} as const;

export function isPrintStatement(item: unknown): item is PrintStatement {
    return reflection.isInstance(item, PrintStatement.$type);
}

export interface ReturnStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'ReturnStatement';
    value?: Expression;
}

export const ReturnStatement = {
    $type: 'ReturnStatement',
    value: 'value'
} as const;

export function isReturnStatement(item: unknown): item is ReturnStatement {
    return reflection.isInstance(item, ReturnStatement.$type);
}

export type Statement = Expression | ExpressionBlock | ForStatement | IfStatement | PrintStatement | ReturnStatement | WhileStatement;

export const Statement = {
    $type: 'Statement'
} as const;

export function isStatement(item: unknown): item is Statement {
    return reflection.isInstance(item, Statement.$type);
}

export interface StringExpression extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | PrintStatement | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'StringExpression';
    value: string;
}

export const StringExpression = {
    $type: 'StringExpression',
    value: 'value'
} as const;

export function isStringExpression(item: unknown): item is StringExpression {
    return reflection.isInstance(item, StringExpression.$type);
}

export type Type = FunctionDeclaration | NamedElement;

export const Type = {
    $type: 'Type'
} as const;

export function isType(item: unknown): item is Type {
    return reflection.isInstance(item, Type.$type);
}

export interface UnaryExpression extends langium.AstNode {
    readonly $container: BinaryExpression | ExpressionBlock | ForStatement | IfStatement | ImplementationModel | MemberCall | ReturnStatement | UnaryExpression | VariableDeclaration | WhileStatement;
    readonly $type: 'UnaryExpression';
    operator: '!' | '+' | '-';
    value: Expression;
}

export const UnaryExpression = {
    $type: 'UnaryExpression',
    operator: 'operator',
    value: 'value'
} as const;

export function isUnaryExpression(item: unknown): item is UnaryExpression {
    return reflection.isInstance(item, UnaryExpression.$type);
}

export interface VariableDeclaration extends langium.AstNode {
    readonly $container: DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'VariableDeclaration';
    array: boolean;
    arrayOfArray: boolean;
    assignment: boolean;
    name: string;
    type: langium.Reference<Class>;
    value?: ConstructorCall | Expression;
}

export const VariableDeclaration = {
    $type: 'VariableDeclaration',
    array: 'array',
    arrayOfArray: 'arrayOfArray',
    assignment: 'assignment',
    name: 'name',
    type: 'type',
    value: 'value'
} as const;

export function isVariableDeclaration(item: unknown): item is VariableDeclaration {
    return reflection.isInstance(item, VariableDeclaration.$type);
}

export interface WhileStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'WhileStatement';
    block: ExpressionBlock;
    condition: Expression;
}

export const WhileStatement = {
    $type: 'WhileStatement',
    block: 'block',
    condition: 'condition'
} as const;